add Settings page and integrate SettingsPanel for user preferences
