+++ frontend/src/pages/Profile.tsx
@@ -7,28 +7,20 @@ import {
   Badge,
   Button,
   Heading,
-  Separator,
   Box,
   TextField,
   TextArea,
-  Switch,
-  Dialog,
   Grid,
   Tabs,
 } from "@radix-ui/themes";
 import {
   CheckCircledIcon,
   Crosshair1Icon,
-  LockClosedIcon,
   Pencil1Icon,
   CheckIcon,
   Cross2Icon,
-  ExitIcon,
 } from "@radix-ui/react-icons";
 import {
-  UserSettings,
-  PasswordChangeForm,
-  AccountDeletionForm,
   SocialLinks,
   RatingDetailed,
   TimeBankResponse,
@@ -42,7 +34,6 @@ import { RatingStars } from "@/components/ui/RatingStars";
 import { InterestSelector } from "@/components/ui/InterestSelector";
 import { InterestChip } from "@/components/ui/InterestChip";
 import { MapLocationPicker } from "@/components/ui/MapLocationPicker";
-import { ConfirmDialog } from "@/components/ui/ConfirmDialog";
 import { PROFILE_PICTURE_PRESETS } from "@/constants/profilePicturePresets";
 import {
   Linkedin,
@@ -83,31 +74,10 @@ export function Profile() {
     interests: [] as string[],
   });
   const [showInterestSelector, setShowInterestSelector] = useState(false);
-  const [settings, setSettings] = useState<UserSettings>({
-    profile_visible: true,
-    show_email: false,
-    show_location: true,
-    email_notifications: true,
-    service_matches_notifications: true,
-    messages_notifications: true,
-  });
-  const [showPasswordDialog, setShowPasswordDialog] = useState(false);
-  const [showDeleteDialog, setShowDeleteDialog] = useState(false);
   const [profilePictureUploading, setProfilePictureUploading] = useState(false);
   const [profilePictureError, setProfilePictureError] = useState<string | null>(
     null,
   );
-  const [passwordForm, setPasswordForm] = useState<PasswordChangeForm>({
-    current_password: "",
-    new_password: "",
-    confirm_password: "",
-  });
-  const [deleteForm, setDeleteForm] = useState<AccountDeletionForm>({
-    password: "",
-  });
-  const [settingsLoading, setSettingsLoading] = useState(false);
-  const [passwordLoading, setPasswordLoading] = useState(false);
-  const [deleteLoading, setDeleteLoading] = useState(false);
   const [myservicesCounts, setMyservicesCounts] = useState({
     services: 0,
     applications: 0,
@@ -199,14 +169,6 @@ export function Profile() {
         portfolio: user.social_links?.portfolio || "",
       },
       interests: user.interests || [],
-    });
-    setSettings({
-      profile_visible: user.profile_visible ?? true,
-      show_email: user.show_email ?? false,
-      show_location: user.show_location ?? true,
-      email_notifications: user.email_notifications ?? true,
-      service_matches_notifications: user.service_matches_notifications ?? true,
-      messages_notifications: user.messages_notifications ?? true,
     });
   }, [user]);
 
@@ -359,108 +321,6 @@ export function Profile() {
     });
   };
 
-  const handleSettingsChange = async (
-    field: keyof UserSettings,
-    value: boolean,
-  ) => {
-    const oldSettings = { ...settings };
-    const newSettings = { ...settings, [field]: value };
-    setSettings(newSettings);
-    setSettingsLoading(true);
-
-    try {
-      const response = await usersApi.updateSettings({ [field]: value });
-      refetchUser();
-      // Update settings from response
-      setSettings({
-        profile_visible: response.data.profile_visible ?? true,
-        show_email: response.data.show_email ?? false,
-        show_location: response.data.show_location ?? true,
-        email_notifications: response.data.email_notifications ?? true,
-        service_matches_notifications:
-          response.data.service_matches_notifications ?? true,
-        messages_notifications: response.data.messages_notifications ?? true,
-      });
-    } catch (error) {
-      console.error("Error updating settings:", error);
-      // Revert on error
-      setSettings(oldSettings);
-      alert("Failed to update settings. Please try again.");
-    } finally {
-      setSettingsLoading(false);
-    }
-  };
-
-  const handlePasswordChange = async () => {
-    if (passwordForm.new_password !== passwordForm.confirm_password) {
-      alert("New passwords do not match");
-      return;
-    }
-
-    if (passwordForm.new_password.length < 8) {
-      alert("Password must be at least 8 characters long");
-      return;
-    }
-
-    setPasswordLoading(true);
-    try {
-      await usersApi.changePassword(passwordForm);
-      // alert("Password changed successfully");
-      setShowPasswordDialog(false);
-      setPasswordForm({
-        current_password: "",
-        new_password: "",
-        confirm_password: "",
-      });
-    } catch (error: any) {
-      console.error("Error changing password:", error);
-      alert(
-        error.response?.data?.detail ||
-          "Failed to change password. Please check your current password.",
-      );
-    } finally {
-      setPasswordLoading(false);
-    }
-  };
-
-  const [deleteConfirmOpen, setDeleteConfirmOpen] = useState(false);
-  const [logoutConfirmOpen, setLogoutConfirmOpen] = useState(false);
-
-  const handleDeleteAccount = async () => {
-    if (!deleteForm.password) {
-      alert("Please enter your password to confirm account deletion");
-      return;
-    }
-    setDeleteConfirmOpen(true);
-  };
-
-  const handleConfirmDeleteAccount = async () => {
-    setDeleteLoading(true);
-    try {
-      await usersApi.deleteAccount(deleteForm);
-      // alert("Account deleted successfully");
-      localStorage.removeItem("access_token");
-      window.location.href = "/";
-    } catch (error: any) {
-      console.error("Error deleting account:", error);
-      alert(
-        error.response?.data?.detail ||
-          "Failed to delete account. Please check your password.",
-      );
-    } finally {
-      setDeleteLoading(false);
-    }
-  };
-
-  const handleLogout = () => {
-    setLogoutConfirmOpen(true);
-  };
-
-  const handleConfirmLogout = () => {
-    localStorage.removeItem("access_token");
-    navigate("/");
-    window.location.reload();
-  };
 
   if (isLoading) {
     return (
@@ -557,27 +417,26 @@ export function Profile() {
           {/* ── Profile Tab ── */}
           <Tabs.Content value="profile">
             <div className="grid gap-6">
-              <Grid columns={{ initial: "1", md: "2" }} gap="6">
-                {/* Profile Information */}
-                <Box>
-                  <Card size="4" className="p-6">
-                    <Flex
+              {/* Profile Information */}
+              <Box>
+                <Card size="4" className="p-6">
+                  <Flex
                       align="center"
                       justify="between"
                       className="col-span-2 mb-4"
-                    >
-                      <Heading size="5">Profile Information</Heading>
-                      {!isEditing ? (
+                  >
+                    <Heading size="5">Profile Information</Heading>
+                    {!isEditing ? (
                         <Button onClick={handleEdit} size="2">
                           <Pencil1Icon className="w-4 h-4 mr-2" />
                           Edit Profile
                         </Button>
-                      ) : (
+                    ) : (
                         <Flex gap="2">
                           <Button
-                            onClick={handleCancel}
-                            variant="soft"
-                            size="2"
+                              onClick={handleCancel}
+                              variant="soft"
+                              size="2"
                           >
                             <Cross2Icon className="w-4 h-4 mr-2" />
                             Cancel
@@ -587,114 +446,116 @@ export function Profile() {
                             Save Changes
                           </Button>
                         </Flex>
-                      )}
-                    </Flex>
+                    )}
+                  </Flex>
 
-                    <div className="space-y-4">
-                      {/* Avatar and Basic Info */}
-                      <Flex align="center" gap="4">
-                        <Avatar
-                          src={
-                            isEditing
+                  {/* Avatar and Basic Info */}
+                  <Flex align="center" gap="4">
+                    <Avatar
+                        src={
+                          isEditing
                               ? getImageUrl(editForm.profile_picture) ||
-                                undefined
+                              undefined
                               : getImageUrl(user.profile_picture) || undefined
-                          }
-                          fallback={user.full_name?.[0] || user.username[0]}
-                          size="6"
-                        />
-                        <div className="flex-1">
-                          <Flex align="center" gap="2" mb="1">
-                            <Heading size="4">
-                              {isEditing ? (
-                                <TextField.Root
+                        }
+                        fallback={user.full_name?.[0] || user.username[0]}
+                        size="6"
+                    />
+                    <div className="flex-1">
+                      <Flex align="center" gap="2" mb="1">
+                        <Heading size="4">
+                          {isEditing ? (
+                              <TextField.Root
                                   value={editForm.full_name}
                                   onChange={(e) =>
-                                    handleInputChange(
-                                      "full_name",
-                                      e.target.value,
-                                    )
+                                      handleInputChange(
+                                          "full_name",
+                                          e.target.value,
+                                      )
                                   }
                                   placeholder="Full Name"
                                   size="2"
-                                />
-                              ) : (
-                                user.full_name || user.username
-                              )}
-                            </Heading>
-                            {user.is_verified && (
-                              <CheckCircledIcon className="w-4 h-4 text-green-600" />
-                            )}
-                          </Flex>
-                          <Text size="3" color="gray">
-                            @{user.username}
-                          </Text>
-                          {user.is_verified && (
-                            <Badge
+                              />
+                          ) : (
+                              user.full_name || user.username
+                          )}
+                        </Heading>
+                        {user.is_verified && (
+                            <CheckCircledIcon className="w-4 h-4 text-green-600" />
+                        )}
+                      </Flex>
+                      <Text size="3" color="gray">
+                        @{user.username}
+                      </Text>
+                      {user.is_verified && (
+                          <Badge
                               color="green"
                               variant="soft"
                               size="1"
                               className="ml-2"
-                            >
-                              Verified User
-                            </Badge>
-                          )}
-                        </div>
-                      </Flex>
+                          >
+                            Verified User
+                          </Badge>
+                      )}
+                    </div>
+                  </Flex>
 
-                      {/* Profile Picture: presets, upload, or URL */}
-                      {isEditing && (
-                        <div className="space-y-2">
-                          <Text size="2" weight="bold" className="block">
-                            Profile Picture
+
+
+
+                  {/* Profile Picture: presets, upload, or URL */}
+                  {isEditing && (
+                      <div className="space-y-2">
+                        <Text size="2" weight="bold" className="block">
+                          Profile Picture
+                        </Text>
+
+                        {/* Preset avatars */}
+                        <div>
+                          <Text size="1" color="gray" className="block mb-2">
+                            Choose a preset avatar
                           </Text>
-
-                          {/* Preset avatars */}
-                          <div>
-                            <Text size="1" color="gray" className="block mb-2">
-                              Choose a preset avatar
-                            </Text>
-                            <Grid columns="6" gap="2" className="max-w-md">
-                              {PROFILE_PICTURE_PRESETS.map((preset) => {
-                                const isSelected =
+                          <Grid columns="6" gap="2" className="max-w-md">
+                            {PROFILE_PICTURE_PRESETS.map((preset) => {
+                              const isSelected =
                                   editForm.profile_picture === preset.url;
-                                return (
+                              return (
                                   <button
-                                    key={preset.id}
-                                    type="button"
-                                    onClick={() => {
-                                      handleInputChange(
-                                        "profile_picture",
-                                        preset.url,
-                                      );
-                                      setProfilePictureError(null);
-                                    }}
-                                    className={`rounded-full p-0.5 transition-all focus:outline-none focus:ring-2 focus:ring-cyan-9 ${
-                                      isSelected
-                                        ? "ring-2 ring-cyan-9 ring-offset-2 ring-offset-gray-1 dark:ring-offset-gray-2"
-                                        : "hover:opacity-90"
-                                    }`}
-                                    title={preset.name}
+                                      key={preset.id}
+                                      type="button"
+                                      onClick={() => {
+                                        handleInputChange(
+                                            "profile_picture",
+                                            preset.url,
+                                        );
+                                        setProfilePictureError(null);
+                                      }}
+                                      className={`rounded-full p-0.5 transition-all focus:outline-none focus:ring-2 focus:ring-cyan-9 ${
+                                          isSelected
+                                              ? "ring-2 ring-cyan-9 ring-offset-2 ring-offset-gray-1 dark:ring-offset-gray-2"
+                                              : "hover:opacity-90"
+                                      }`}
+                                      title={preset.name}
                                   >
                                     <Avatar
-                                      src={preset.url}
-                                      fallback={preset.name[0]}
-                                      size="3"
-                                      radius="full"
-                                      className="w-full aspect-square"
+                                        src={preset.url}
+                                        fallback={preset.name[0]}
+                                        size="3"
+                                        radius="full"
+                                        className="w-full aspect-square"
                                     />
                                   </button>
-                                );
-                              })}
-                            </Grid>
-                          </div>
+                              );
+                            })}
+                          </Grid>
+                        </div>
 
-                          <Text size="1" color="gray" className="block mt-6">
-                            Or upload your own:
-                          </Text>
-                          <Flex gap="2" align="center" wrap="wrap">
-                            <label className="cursor-pointer">
-                              <input
+                        <Text size="1" color="gray" className="block mt-6">
+                          Or upload your own:
+                        </Text>
+                        <Flex gap="2" align="center" wrap="wrap">
+                          <label className="cursor-pointer">
+                            <input
                                 type="file"
                                 accept="image/jpeg,image/png,image/webp"
                                 className="sr-only"
@@ -705,7 +566,7 @@ export function Profile() {
                                   const maxMb = 5;
                                   if (file.size > maxMb * 1024 * 1024) {
                                     setProfilePictureError(
-                                      `File must be under ${maxMb} MB`,
+                                        `File must be under ${maxMb} MB`,
                                     );
                                     return;
                                   }
@@ -713,16 +574,16 @@ export function Profile() {
                                   setProfilePictureUploading(true);
                                   try {
                                     const res =
-                                      await uploadApi.uploadProfilePicture(
-                                        file,
-                                      );
+                                        await uploadApi.uploadProfilePicture(
+                                            file,
+                                        );
                                     handleInputChange(
-                                      "profile_picture",
-                                      res.data.url,
+                                        "profile_picture",
+                                        res.data.url,
                                     );
                                   } catch (err: any) {
                                     setProfilePictureError(
-                                      err.response?.data?.detail ||
+                                        err.response?.data?.detail ||
                                         "Upload failed",
                                     );
                                   } finally {
@@ -730,620 +591,437 @@ export function Profile() {
                                     e.target.value = "";
                                   }
                                 }}
-                              />
-                              <Button
+                            />
+                            <Button
                                 type="button"
                                 size="2"
                                 variant="soft"
                                 asChild
-                              >
+                            >
                                 <span>
                                   {profilePictureUploading
-                                    ? "Uploading..."
-                                    : "Upload photo"}
+                                      ? "Uploading..."
+                                      : "Upload photo"}
                                 </span>
-                              </Button>
-                            </label>
-                            <Text size="2" color="gray">
-                              or paste URL:
-                            </Text>
-                          </Flex>
-                          <TextField.Root
+                            </Button>
+                          </label>
+                          <Text size="2" color="gray">
+                            or paste URL:
+                          </Text>
+                        </Flex>
+                        <TextField.Root
                             value={editForm.profile_picture}
                             onChange={(e) => {
                               handleInputChange(
-                                "profile_picture",
-                                e.target.value,
+                                  "profile_picture",
+                                  e.target.value,
                               );
                               setProfilePictureError(null);
                             }}
                             placeholder="https://example.com/photo.jpg or /uploads/..."
                             size="2"
-                          />
-                          {profilePictureError && (
+                        />
+                        {profilePictureError && (
                             <Text size="2" color="red">
                               {profilePictureError}
                             </Text>
-                          )}
-                        </div>
-                      )}
+                        )}
+                      </div>
+                  )}
 
-                      {/* Email */}
-                      <div>
-                        <Text size="2" weight="bold" mr="2">
-                          Email
-                        </Text>
-                        {isEditing ? (
+
+                  <div className="grid grid-cols-2 gap-3">
+                    {/* Email */}
+                    <div>
+                      <Text size="2" weight="bold" mr="2">
+                        Email
+                      </Text>
+                      {isEditing ? (
                           <TextField.Root
-                            value={editForm.email}
-                            onChange={(e) =>
-                              handleInputChange("email", e.target.value)
-                            }
-                            placeholder="Email"
-                            size="2"
+                              value={editForm.email}
+                              onChange={(e) =>
+                                  handleInputChange("email", e.target.value)
+                              }
+                              placeholder="Email"
+                              size="2"
                           />
-                        ) : (
+                      ) : (
                           <Text size="2">{user.email}</Text>
-                        )}
-                      </div>
+                      )}
+                    </div>
 
-                      {/* Bio */}
-                      <div>
-                        <Text size="2" weight="bold" mr="2">
-                          Bio
-                        </Text>
-                        {isEditing ? (
+                    {/* Bio */}
+                    <div>
+                      <Text size="2" weight="bold" mr="2">
+                        Bio
+                      </Text>
+                      {isEditing ? (
                           <TextArea
-                            value={editForm.bio}
-                            onChange={(e) =>
-                              handleInputChange("bio", e.target.value)
-                            }
-                            placeholder="Tell us about yourself..."
-                            size="2"
-                            rows={3}
+                              value={editForm.bio}
+                              onChange={(e) =>
+                                  handleInputChange("bio", e.target.value)
+                              }
+                              placeholder="Tell us about yourself..."
+                              size="2"
+                              rows={3}
                           />
-                        ) : (
+                      ) : (
                           <Text size="2">{user.bio || "No bio provided"}</Text>
-                        )}
-                      </div>
+                      )}
+                    </div>
 
-                      {/* Location */}
-                      <div>
-                        <Text size="2" weight="bold" mr="2">
-                          Location
-                        </Text>
-                        {isEditing ? (
+                    {/* Location */}
+                    <div>
+                      <Text size="2" weight="bold" mr="2">
+                        Location
+                      </Text>
+                      {isEditing ? (
                           <MapLocationPicker
-                            value={editForm.location}
-                            onChange={(loc) =>
-                              setEditForm((prev) => ({
-                                ...prev,
-                                location: {
-                                  latitude: loc.latitude,
-                                  longitude: loc.longitude,
-                                  address: loc.address || "",
-                                },
-                              }))
-                            }
-                            height={180}
-                            markerColor="#2563eb"
+                              value={editForm.location}
+                              onChange={(loc) =>
+                                  setEditForm((prev) => ({
+                                    ...prev,
+                                    location: {
+                                      latitude: loc.latitude,
+                                      longitude: loc.longitude,
+                                      address: loc.address || "",
+                                    },
+                                  }))
+                              }
+                              height={180}
+                              markerColor="#2563eb"
                           />
-                        ) : (
+                      ) : (
                           <Flex align="center" gap="2">
                             <Crosshair1Icon className="w-4 h-4" />
                             <Text size="2">
                               {typeof user.location === "string"
-                                ? user.location || "No location provided"
-                                : (user.location as any)?.address ||
+                                  ? user.location || "No location provided"
+                                  : (user.location as any)?.address ||
                                   "No location provided"}
                             </Text>
                           </Flex>
-                        )}
-                      </div>
+                      )}
+                    </div>
 
-                      {/* Average Rating */}
-                      <div>
-                        <Text size="2" weight="bold" className="block mb-1">
-                          Average Rating
-                        </Text>
-                        {ratingCount > 0 && averageRating != null ? (
+                    {/* Average Rating */}
+                    <div>
+                      <Text size="2" weight="bold" className="block mb-1">
+                        Average Rating
+                      </Text>
+                      {ratingCount > 0 && averageRating != null ? (
                           <Flex align="center" gap="2">
                             <RatingStars
-                              value={Math.round(averageRating * 10) / 10}
-                              readonly
-                              size={18}
+                                value={Math.round(averageRating * 10) / 10}
+                                readonly
+                                size={18}
                             />
                             <Text size="2" color="gray">
                               {averageRating.toFixed(1)} ({ratingCount} rating
                               {ratingCount !== 1 ? "s" : ""})
                             </Text>
                           </Flex>
-                        ) : (
+                      ) : (
                           <Text size="2" color="gray">
                             No ratings yet
                           </Text>
-                        )}
-                      </div>
-
-                      <Separator />
-
-                      {/* Social Links */}
-                      <div>
-                        <Text size="2" weight="bold" className="block mb-2">
-                          Social Links
-                        </Text>
-                        {isEditing ? (
+                      )}
+                    </div>
+                    {/* Social Links */}
+                    <div>
+                      <Text size="2" weight="bold" className="block mb-2">
+                        Social Links
+                      </Text>
+                      {isEditing ? (
                           <div className="space-y-2">
                             {(
-                              [
                                 [
-                                  "linkedin",
-                                  "LinkedIn",
-                                  "https://linkedin.com/in/...",
-                                ],
-                                ["github", "GitHub", "https://github.com/..."],
-                                [
-                                  "twitter",
-                                  "Twitter / X",
-                                  "https://twitter.com/...",
-                                ],
-                                [
-                                  "instagram",
-                                  "Instagram",
-                                  "https://instagram.com/...",
-                                ],
-                                ["website", "Website", "https://yoursite.com"],
-                                [
-                                  "portfolio",
-                                  "Portfolio",
-                                  "https://portfolio.com",
-                                ],
-                              ] as const
+                                  [
+                                    "linkedin",
+                                    "LinkedIn",
+                                    "https://linkedin.com/in/...",
+                                  ],
+                                  ["github", "GitHub", "https://github.com/..."],
+                                  [
+                                    "twitter",
+                                    "Twitter / X",
+                                    "https://twitter.com/...",
+                                  ],
+                                  [
+                                    "instagram",
+                                    "Instagram",
+                                    "https://instagram.com/...",
+                                  ],
+                                  ["website", "Website", "https://yoursite.com"],
+                                  [
+                                    "portfolio",
+                                    "Portfolio",
+                                    "https://portfolio.com",
+                                  ],
+                                ] as const
                             ).map(([key, label, placeholder]) => (
-                              <div key={key}>
-                                <Text
-                                  size="1"
-                                  color="gray"
-                                  className="block mb-0.5"
-                                >
-                                  {label}
-                                </Text>
-                                <TextField.Root
-                                  value={
-                                    (editForm.social_links as any)[key] || ""
-                                  }
-                                  onChange={(e) =>
-                                    handleSocialLinkChange(
-                                      key as keyof SocialLinks,
-                                      e.target.value,
-                                    )
-                                  }
-                                  placeholder={placeholder}
-                                  size="1"
-                                />
-                              </div>
+                                <div key={key}>
+                                  <Text
+                                      size="1"
+                                      color="gray"
+                                      className="block mb-0.5"
+                                  >
+                                    {label}
+                                  </Text>
+                                  <TextField.Root
+                                      value={
+                                          (editForm.social_links as any)[key] || ""
+                                      }
+                                      onChange={(e) =>
+                                          handleSocialLinkChange(
+                                              key as keyof SocialLinks,
+                                              e.target.value,
+                                          )
+                                      }
+                                      placeholder={placeholder}
+                                      size="1"
+                                  />
+                                </div>
                             ))}
                           </div>
-                        ) : (
+                      ) : (
                           <Flex gap="3" wrap="wrap">
                             {user.social_links?.linkedin && (
-                              <a
-                                href={user.social_links.linkedin}
-                                target="_blank"
-                                rel="noopener noreferrer"
-                                title="LinkedIn"
-                              >
-                                <Linkedin
-                                  size={20}
-                                  className="text-gray-600 hover:text-blue-600 transition-colors"
-                                />
-                              </a>
+                                <a
+                                    href={user.social_links.linkedin}
+                                    target="_blank"
+                                    rel="noopener noreferrer"
+                                    title="LinkedIn"
+                                >
+                                  <Linkedin
+                                      size={20}
+                                      className="text-gray-600 hover:text-blue-600 transition-colors"
+                                  />
+                                </a>
                             )}
                             {user.social_links?.github && (
-                              <a
-                                href={user.social_links.github}
-                                target="_blank"
-                                rel="noopener noreferrer"
-                                title="GitHub"
-                              >
-                                <Github
-                                  size={20}
-                                  className="text-gray-600 hover:text-gray-900 transition-colors"
-                                />
-                              </a>
+                                <a
+                                    href={user.social_links.github}
+                                    target="_blank"
+                                    rel="noopener noreferrer"
+                                    title="GitHub"
+                                >
+                                  <Github
+                                      size={20}
+                                      className="text-gray-600 hover:text-gray-900 transition-colors"
+                                  />
+                                </a>
                             )}
                             {user.social_links?.twitter && (
-                              <a
-                                href={user.social_links.twitter}
-                                target="_blank"
-                                rel="noopener noreferrer"
-                                title="Twitter / X"
-                              >
-                                <Twitter
-                                  size={20}
-                                  className="text-gray-600 hover:text-sky-500 transition-colors"
-                                />
-                              </a>
+                                <a
+                                    href={user.social_links.twitter}
+                                    target="_blank"
+                                    rel="noopener noreferrer"
+                                    title="Twitter / X"
+                                >
+                                  <Twitter
+                                      size={20}
+                                      className="text-gray-600 hover:text-sky-500 transition-colors"
+                                  />
+                                </a>
                             )}
                             {user.social_links?.instagram && (
-                              <a
-                                href={user.social_links.instagram}
-                                target="_blank"
-                                rel="noopener noreferrer"
-                                title="Instagram"
-                              >
-                                <Instagram
-                                  size={20}
-                                  className="text-gray-600 hover:text-pink-500 transition-colors"
-                                />
-                              </a>
+                                <a
+                                    href={user.social_links.instagram}
+                                    target="_blank"
+                                    rel="noopener noreferrer"
+                                    title="Instagram"
+                                >
+                                  <Instagram
+                                      size={20}
+                                      className="text-gray-600 hover:text-pink-500 transition-colors"
+                                  />
+                                </a>
                             )}
                             {user.social_links?.website && (
-                              <a
-                                href={user.social_links.website}
-                                target="_blank"
-                                rel="noopener noreferrer"
-                                title="Website"
-                              >
-                                <Globe
-                                  size={20}
-                                  className="text-gray-600 hover:text-green-600 transition-colors"
-                                />
-                              </a>
+                                <a
+                                    href={user.social_links.website}
+                                    target="_blank"
+                                    rel="noopener noreferrer"
+                                    title="Website"
+                                >
+                                  <Globe
+                                      size={20}
+                                      className="text-gray-600 hover:text-green-600 transition-colors"
+                                  />
+                                </a>
                             )}
                             {user.social_links?.portfolio && (
-                              <a
-                                href={user.social_links.portfolio}
-                                target="_blank"
-                                rel="noopener noreferrer"
-                                title="Portfolio"
-                              >
-                                <Briefcase
-                                  size={20}
-                                  className="text-gray-600 hover:text-amber-600 transition-colors"
-                                />
-                              </a>
+                                <a
+                                    href={user.social_links.portfolio}
+                                    target="_blank"
+                                    rel="noopener noreferrer"
+                                    title="Portfolio"
+                                >
+                                  <Briefcase
+                                      size={20}
+                                      className="text-gray-600 hover:text-amber-600 transition-colors"
+                                  />
+                                </a>
                             )}
                             {!user.social_links?.linkedin &&
-                              !user.social_links?.github &&
-                              !user.social_links?.twitter &&
-                              !user.social_links?.instagram &&
-                              !user.social_links?.website &&
-                              !user.social_links?.portfolio && (
-                                <Text size="2" color="gray">
-                                  No social links added
-                                </Text>
-                              )}
+                                !user.social_links?.github &&
+                                !user.social_links?.twitter &&
+                                !user.social_links?.instagram &&
+                                !user.social_links?.website &&
+                                !user.social_links?.portfolio && (
+                                    <Text size="2" color="gray">
+                                      No social links added
+                                    </Text>
+                                )}
                           </Flex>
-                        )}
-                      </div>
+                      )}
+                    </div>
 
-                      <Separator />
+                    
 
-                      {/* Interests */}
-                      <div>
-                        <Flex justify="between" align="center" mb="3">
-                          <Text size="2" weight="bold">
-                            Interests
-                          </Text>
-                          <Button
+                    {/* Interests */}
+                    <div>
+                      <Flex justify="between" align="center" mb="3">
+                        <Text size="2" weight="bold">
+                          Interests
+                        </Text>
+                        <Button
                             size="1"
                             variant="soft"
                             color="lime"
                             className="rounded-full"
                             onClick={() => setShowInterestSelector(true)}
-                          >
-                            {(user.interests?.length || 0) > 0
+                        >
+                          {(user.interests?.length || 0) > 0
                               ? "Update Interests"
                               : "Add Interests"}
-                          </Button>
-                        </Flex>
-                        {(user.interests?.length || 0) > 0 ? (
+                        </Button>
+                      </Flex>
+                      {(user.interests?.length || 0) > 0 ? (
                           <Flex gap="2" wrap="wrap">
                             {user.interests!.map((interest) => (
-                              <InterestChip
-                                key={interest}
-                                name={interest}
-                                size="sm"
-                                showIcon
-                              />
+                                <InterestChip
+                                    key={interest}
+                                    name={interest}
+                                    size="sm"
+                                    showIcon
+                                />
                             ))}
                           </Flex>
-                        ) : (
+                      ) : (
                           <button
-                            type="button"
-                            onClick={() => setShowInterestSelector(true)}
-                            className="rounded-xl border-2 border-dashed px-4 py-3 text-left text-sm transition-colors"
-                            style={{
-                              borderColor: "var(--gray-6)",
-                              backgroundColor: "var(--gray-1)",
-                              color: "var(--gray-10)",
-                            }}
-                            onMouseEnter={(e) => {
-                              e.currentTarget.style.borderColor =
-                                "var(--lime-6)";
-                              e.currentTarget.style.backgroundColor =
-                                "var(--lime-2)";
-                              e.currentTarget.style.color = "var(--lime-11)";
-                            }}
-                            onMouseLeave={(e) => {
-                              e.currentTarget.style.borderColor =
-                                "var(--gray-6)";
-                              e.currentTarget.style.backgroundColor =
-                                "var(--gray-1)";
-                              e.currentTarget.style.color = "var(--gray-10)";
-                            }}
+                              type="button"
+                              onClick={() => setShowInterestSelector(true)}
+                              className="rounded-xl border-2 border-dashed px-4 py-3 text-left text-sm transition-colors"
+                              style={{
+                                borderColor: "var(--gray-6)",
+                                backgroundColor: "var(--gray-1)",
+                                color: "var(--gray-10)",
+                              }}
+                              onMouseEnter={(e) => {
+                                e.currentTarget.style.borderColor =
+                                    "var(--lime-6)";
+                                e.currentTarget.style.backgroundColor =
+                                    "var(--lime-2)";
+                                e.currentTarget.style.color = "var(--lime-11)";
+                              }}
+                              onMouseLeave={(e) => {
+                                e.currentTarget.style.borderColor =
+                                    "var(--gray-6)";
+                                e.currentTarget.style.backgroundColor =
+                                    "var(--gray-1)";
+                                e.currentTarget.style.color = "var(--gray-10)";
+                              }}
                           >
                             Add interests to help others discover you
                           </button>
-                        )}
-                      </div>
+                      )}
+                    </div>
 
-                      <Separator />
+                    
 
-                      {/* Communities */}
-                      {(userCommunitiesLoading || userCommunities.length > 0) && (
+                    {/* Communities */}
+                    {(userCommunitiesLoading || userCommunities.length > 0) && (
                         <div>
                           <Flex align="center" gap="2" className="mb-2">
                             <Text size="2" weight="bold">
                               Communities
                             </Text>
                             {!userCommunitiesLoading && (
-                              <Text size="1" color="gray">
-                                {userCommunities.length}
-                              </Text>
+                                <Text size="1" color="gray">
+                                  {userCommunities.length}
+                                </Text>
                             )}
                           </Flex>
                           {userCommunitiesLoading ? (
-                            <Text size="1" color="gray">
-                              Loading communities...
-                            </Text>
+                              <Text size="1" color="gray">
+                                Loading communities...
+                              </Text>
                           ) : (
-                            <Flex gap="2" wrap="wrap">
-                              {userCommunities.slice(0, 12).map((community) => (
-                                <button
-                                  key={community._id}
-                                  type="button"
-                                  title={community.name}
-                                  aria-label={`Open ${community.name}`}
-                                  onClick={() =>
-                                    navigate(`/forum/communities/${community._id}`)
-                                  }
-                                  className="rounded-full ring-1 ring-[var(--gray-6)] transition hover:scale-105 focus:outline-none focus:ring-2 focus:ring-[var(--grass-8)]"
-                                >
-                                  <Avatar
-                                    size="3"
-                                    src={getImageUrl(community.avatar_url)}
-                                    fallback={community.name[0]}
-                                    radius="full"
-                                  />
-                                </button>
-                              ))}
-                              {userCommunities.length > 12 && (
-                                <span
-                                  title={`${userCommunities.length - 12} more communities`}
-                                  className="flex h-10 w-10 items-center justify-center rounded-full bg-[var(--gray-3)] text-sm font-medium text-[var(--gray-11)] ring-1 ring-[var(--gray-6)]"
-                                >
+                              <Flex gap="2" wrap="wrap">
+                                {userCommunities.slice(0, 12).map((community) => (
+                                    <button
+                                        key={community._id}
+                                        type="button"
+                                        title={community.name}
+                                        aria-label={`Open ${community.name}`}
+                                        onClick={() =>
+                                            navigate(`/forum/communities/${community._id}`)
+                                        }
+                                        className="rounded-full ring-1 ring-[var(--gray-6)] transition hover:scale-105 focus:outline-none focus:ring-2 focus:ring-[var(--grass-8)]"
+                                    >
+                                      <Avatar
+                                          size="3"
+                                          src={getImageUrl(community.avatar_url)}
+                                          fallback={community.name[0]}
+                                          radius="full"
+                                      />
+                                    </button>
+                                ))}
+                                {userCommunities.length > 12 && (
+                                    <span
+                                        title={`${userCommunities.length - 12} more communities`}
+                                        className="flex h-10 w-10 items-center justify-center rounded-full bg-[var(--gray-3)] text-sm font-medium text-[var(--gray-11)] ring-1 ring-[var(--gray-6)]"
+                                    >
                                   +{userCommunities.length - 12}
                                 </span>
-                              )}
-                            </Flex>
+                                )}
+                              </Flex>
                           )}
                         </div>
-                      )}
+                    )}
 
-                      <Separator />
+                    
 
-                      {/* Stats */}
-                      <div className="space-y-2">
-                        <Flex justify="between" align="center">
-                          <Text size="2">TimeBank Balance</Text>
-                          <Text size="2" weight="bold">
-                            {user.timebank_balance} hours
-                          </Text>
-                        </Flex>
-                        <Flex justify="between" align="center">
-                          <Text size="2">Member Since</Text>
-                          <Text size="2">
-                            {new Date(user.created_at).toLocaleDateString()}
-                          </Text>
-                        </Flex>
-                        <Flex justify="between" align="center">
-                          <Text size="2">Status</Text>
-                          <Badge
+                    {/* Stats */}
+                    <div className="space-y-2">
+                      <Flex justify="between" align="center">
+                        <Text size="2">TimeBank Balance</Text>
+                        <Text size="2" weight="bold">
+                          {user.timebank_balance} hours
+                        </Text>
+                      </Flex>
+                      <Flex justify="between" align="center">
+                        <Text size="2">Member Since</Text>
+                        <Text size="2">
+                          {new Date(user.created_at).toLocaleDateString()}
+                        </Text>
+                      </Flex>
+                      <Flex justify="between" align="center">
+                        <Text size="2">Status</Text>
+                        <Badge
                             color={user.is_active ? "green" : "red"}
                             variant="soft"
-                          >
-                            {user.is_active ? "Active" : "Inactive"}
-                          </Badge>
-                        </Flex>
-                      </div>
+                        >
+                          {user.is_active ? "Active" : "Inactive"}
+                        </Badge>
+                      </Flex>
                     </div>
-                  </Card>
-                </Box>
+                  </div>
+                </Card>
+              </Box>
 
-                {/* Settings */}
-                <Box>
-                  <Card size="4" className="p-6">
-                    <Heading size="5" mb="4">
-                      Settings & Preferences
-                    </Heading>
-
-                    <div className="space-y-6">
-                      {/* Privacy Settings */}
-                      <div>
-                        <Heading size="4" mb="3">
-                          Privacy
-                        </Heading>
-                        <div className="space-y-3">
-                          {/*
-                    <Flex justify="between" align="center">
-                    <div className="grid gap-1">
-                      <Text size="2" weight="bold">
-                        Profile Visibility
-                      </Text>
-                      <Text size="1" color="gray">
-                        Make your profile visible to other users
-                      </Text>
-                    </div>
-                    <Switch
-                      checked={settings.profile_visible}
-                      onCheckedChange={(checked) =>
-                        handleSettingsChange("profile_visible", checked)
-                      }
-                      disabled={settingsLoading}
-                    />
-                  </Flex>
-                    */}
-                          <Flex justify="between" align="center">
-                            <div className="grid gap-1">
-                              <Text size="2" weight="bold">
-                                Show Email
-                              </Text>
-                              <Text size="1" color="gray">
-                                Display your email on your profile
-                              </Text>
-                            </div>
-                            <Switch
-                              checked={settings.show_email}
-                              onCheckedChange={(checked) =>
-                                handleSettingsChange("show_email", checked)
-                              }
-                              disabled={settingsLoading}
-                            />
-                          </Flex>
-                          <Flex justify="between" align="center">
-                            <div className="grid gap-1">
-                              <Text size="2" weight="bold">
-                                Show Location
-                              </Text>
-                              <Text size="1" color="gray">
-                                Display your location on your profile
-                              </Text>
-                            </div>
-                            <Switch
-                              checked={settings.show_location}
-                              onCheckedChange={(checked) =>
-                                handleSettingsChange("show_location", checked)
-                              }
-                              disabled={settingsLoading}
-                            />
-                          </Flex>
-                        </div>
-                      </div>
-
-                      <Separator />
-
-                      {/* Notification Settings */}
-                      <div>
-                        <Heading size="4" mb="3">
-                          Notifications
-                        </Heading>
-                        <div className="space-y-3">
-                          <Flex justify="between" align="center">
-                            <div className="grid gap-1">
-                              <Text size="2" weight="bold">
-                                Email Notifications
-                              </Text>
-                              <Text size="1" color="gray">
-                                Receive notifications via email
-                              </Text>
-                            </div>
-                            <Switch
-                              checked={settings.email_notifications}
-                              onCheckedChange={(checked) =>
-                                handleSettingsChange(
-                                  "email_notifications",
-                                  checked,
-                                )
-                              }
-                              disabled={settingsLoading}
-                            />
-                          </Flex>
-                          <Flex justify="between" align="center">
-                            <div className="grid gap-1">
-                              <Text size="2" weight="bold">
-                                Service Matches
-                              </Text>
-                              <Text size="1" color="gray">
-                                Get notified when services match your needs
-                              </Text>
-                            </div>
-                            <Switch
-                              checked={settings.service_matches_notifications}
-                              onCheckedChange={(checked) =>
-                                handleSettingsChange(
-                                  "service_matches_notifications",
-                                  checked,
-                                )
-                              }
-                              disabled={settingsLoading}
-                            />
-                          </Flex>
-                          <Flex justify="between" align="center">
-                            <div className="grid gap-1">
-                              <Text size="2" weight="bold">
-                                Messages
-                              </Text>
-                              <Text size="1" color="gray">
-                                Get notified about new messages
-                              </Text>
-                            </div>
-                            <Switch
-                              checked={settings.messages_notifications}
-                              onCheckedChange={(checked) =>
-                                handleSettingsChange(
-                                  "messages_notifications",
-                                  checked,
-                                )
-                              }
-                              disabled={settingsLoading}
-                            />
-                          </Flex>
-                        </div>
-                      </div>
-
-                      <Separator />
-
-                      {/* Account Settings */}
-                      <div>
-                        <Heading size="4" mb="3">
-                          Account
-                        </Heading>
-                        <div className="space-y-3">
-                          <Button
-                            variant="soft"
-                            size="2"
-                            className="w-full justify-start"
-                            onClick={() => setShowPasswordDialog(true)}
-                          >
-                            <LockClosedIcon className="w-4 h-4 mr-2" />
-                            Change Password
-                          </Button>
-                          <Button
-                            variant="soft"
-                            size="2"
-                            className="w-full justify-start"
-                            onClick={handleLogout}
-                          >
-                            <ExitIcon className="w-4 h-4 mr-2" />
-                            Logout
-                          </Button>
-                          <Button
-                            variant="soft"
-                            color="red"
-                            size="2"
-                            className="w-full justify-start"
-                            onClick={() => setShowDeleteDialog(true)}
-                          >
-                            Delete Account
-                          </Button>
-                        </div>
-                      </div>
-                    </div>
-                  </Card>
-                </Box>
-              </Grid>
               <BadgeDisplay />
 
               <ActivitySummarySection
@@ -1405,158 +1083,6 @@ export function Profile() {
         onSave={handleInterestsSave}
       />
 
-      {/* Change Password Dialog */}
-      <Dialog.Root
-        open={showPasswordDialog}
-        onOpenChange={setShowPasswordDialog}
-      >
-        <Dialog.Content className="max-w-md">
-          <Dialog.Title>Change Password</Dialog.Title>
-          <Dialog.Description className="mb-4">
-            Enter your current password and choose a new one.
-          </Dialog.Description>
-
-          <div className="space-y-4">
-            <div>
-              <Text size="2" weight="bold" className="block mb-1">
-                Current Password
-              </Text>
-              <TextField.Root
-                type="password"
-                value={passwordForm.current_password}
-                onChange={(e) =>
-                  setPasswordForm({
-                    ...passwordForm,
-                    current_password: e.target.value,
-                  })
-                }
-                placeholder="Enter current password"
-                size="2"
-              />
-            </div>
-
-            <div>
-              <Text size="2" weight="bold" className="block mb-1">
-                New Password
-              </Text>
-              <TextField.Root
-                type="password"
-                value={passwordForm.new_password}
-                onChange={(e) =>
-                  setPasswordForm({
-                    ...passwordForm,
-                    new_password: e.target.value,
-                  })
-                }
-                placeholder="Enter new password"
-                size="2"
-              />
-            </div>
-
-            <div>
-              <Text size="2" weight="bold" className="block mb-1">
-                Confirm New Password
-              </Text>
-              <TextField.Root
-                type="password"
-                value={passwordForm.confirm_password}
-                onChange={(e) =>
-                  setPasswordForm({
-                    ...passwordForm,
-                    confirm_password: e.target.value,
-                  })
-                }
-                placeholder="Confirm new password"
-                size="2"
-              />
-            </div>
-          </div>
-
-          <Flex gap="3" mt="4" justify="end">
-            <Button
-              variant="soft"
-              onClick={() => {
-                setShowPasswordDialog(false);
-                setPasswordForm({
-                  current_password: "",
-                  new_password: "",
-                  confirm_password: "",
-                });
-              }}
-            >
-              Cancel
-            </Button>
-            <Button onClick={handlePasswordChange} disabled={passwordLoading}>
-              {passwordLoading ? "Changing..." : "Change Password"}
-            </Button>
-          </Flex>
-        </Dialog.Content>
-      </Dialog.Root>
-
-      {/* Delete Account Dialog */}
-      <Dialog.Root open={showDeleteDialog} onOpenChange={setShowDeleteDialog}>
-        <Dialog.Content className="max-w-md">
-          <Dialog.Title color="red">Delete Account</Dialog.Title>
-          <Dialog.Description className="mb-4">
-            This action cannot be undone. This will permanently delete your
-            account and all associated data.
-          </Dialog.Description>
-
-          <div className="space-y-4">
-            <div>
-              <Text size="2" weight="bold" className="block mb-1">
-                Enter your password to confirm
-              </Text>
-              <TextField.Root
-                type="password"
-                value={deleteForm.password}
-                onChange={(e) =>
-                  setDeleteForm({ ...deleteForm, password: e.target.value })
-                }
-                placeholder="Enter your password"
-                size="2"
-              />
-            </div>
-          </div>
-
-          <Flex gap="3" mt="4" justify="end">
-            <Button
-              variant="soft"
-              onClick={() => {
-                setShowDeleteDialog(false);
-                setDeleteForm({ password: "" });
-              }}
-            >
-              Cancel
-            </Button>
-            <Button
-              color="red"
-              onClick={handleDeleteAccount}
-              disabled={deleteLoading}
-            >
-              {deleteLoading ? "Deleting..." : "Delete Account"}
-            </Button>
-          </Flex>
-        </Dialog.Content>
-      </Dialog.Root>
-
-      <ConfirmDialog
-        open={deleteConfirmOpen}
-        onOpenChange={setDeleteConfirmOpen}
-        title="Delete your account?"
-        description="Are you sure you want to delete your account? This action cannot be undone."
-        confirmLabel="Delete account"
-        variant="danger"
-        onConfirm={handleConfirmDeleteAccount}
-      />
-      <ConfirmDialog
-        open={logoutConfirmOpen}
-        onOpenChange={setLogoutConfirmOpen}
-        title="Log out?"
-        description="Are you sure you want to logout?"
-        confirmLabel="Log out"
-        onConfirm={handleConfirmLogout}
-      />
       {/* Rejected Requests Dialog
 
        <Dialog.Root
